+++ src/datacache/plugin_datacache_postgres.c
@@ -301,10 +301,10 @@ delete_by_rowid (struct Plugin *plugin,
  * @param discard_time when to discard the value in any case
  * @return 0 on error, number of bytes used otherwise
  */
-static uint32_t 
+static size_t 
 postgres_plugin_put (void *cls,
 		     const GNUNET_HashCode * key,
-		     uint32_t size,
+		     size_t size,
 		     const char *data,
 		     enum GNUNET_BLOCK_Type type,
 		     struct GNUNET_TIME_Absolute discard_time)
